+++ tools/hsbench/common.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2018, Intel Corporation
+ * Copyright (c) 2016-2019, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -41,6 +41,7 @@ extern unsigned int somPrecisionMode;
 extern bool forceEditDistance;
 extern unsigned editDistance;
 extern bool printCompressSize;
+extern bool useLiteralApi;
 
 /** Structure for the result of a single complete scan. */
 struct ResultEntry {