+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 This is a list of notable changes to Hyperscan, in reverse chronological order.
 
+## [5.2.0] 2019-07-12
+- Literal API: add new API `hs_compile_lit()` and `hs_compile_lit_multi()` to
+  process pure literal rule sets. The 2 literal APIs treat each expression text
+  in a literal sense without recognizing any regular grammers.
+- Logical combination: add support for purely negative combinations, which
+  report match at EOD in case of no sub-expressions matched.
+- Windows porting: support shared library (DLL) on Windows with available tools
+  hscheck, hsbench and hsdump.
+- Bugfix for issue #148: fix uninitialized use of `scatter_unit_uX` due to
+  padding.
+- Bugfix for issue #155: fix numerical result out of range error.
+- Bugfix for issue #165: avoid corruption of pending combination report in
+  streaming mode.
+- Bugfix for issue #174: fix scratch free issue when memory allocation fails.
+
 ## [5.1.1] 2019-04-03
 - Add extra detection and handling when invalid rose programs are triggered.
 - Bugfix for issue #136: fix CMake parsing of CPU architecure for GCC-9.